requirements.txt
click==8.1.3
pygithub==1.55
pyyaml==6.0
gitpython